refactor main classes (#17)
* don't ignore the .gem file (so we can check it in to github)

* nevermind - don't include a gem file

* renamed dir (added s on the end)

* version bump

* refactor: pass options to main class that deals with options, then calls the rendering class (get stuff out of the bin file)
