+++ lib/aasm_statecharts.rb
@@ -1,2 +1,3 @@
 require 'aasm-statecharts/version'
 require 'aasm-statecharts/aasm_statecharts'
+require 'aasm-statecharts/chart_renderer'
