+++ spec/spec-out/many_states.dot
@@ -35,7 +35,7 @@ digraph statechart {
 		lp="446.14,171.1",
 		pos="e,184.89,81.698 181.72,264.34 233.01,259.2 312.05,247.79 330.59,225.19 361.1,188.01 362.88,154.36 332.59,117 315.63,96.073 241.28,\
 86.517 195.12,82.531"];
-	"4a8a7b29-e71c-49a9-9a29-791d3adfce20"	 [color=black,
+	"4b3db322-968d-4ad9-b55d-2004a01be505"	 [color=black,
 		fixedsize=true,
 		fontsize=8,
 		height=0.41667,
@@ -43,7 +43,7 @@ digraph statechart {
 		pos="136.59,346.69",
 		shape=doublecircle,
 		width=0.41667];
-	"4a8a7b29-e71c-49a9-9a29-791d3adfce20" -> a	 [pos="e,136.59,294.26 136.59,331.44 136.59,323.76 136.59,313.93 136.59,304.38"];
+	"4b3db322-968d-4ad9-b55d-2004a01be505" -> a	 [pos="e,136.59,294.26 136.59,331.44 136.59,323.76 136.59,313.93 136.59,304.38"];
 	b -> a	 [label=" z  / z1_before(); z2_before(); z1_after(); z2_after(); ",
 		lp="223.63,220.02",
 		pos="e,128.64,243.47 99.981,196.53 105.25,202.21 110.43,208.47 114.59,214.84 118.44,220.74 121.84,227.42 124.74,234"];
@@ -53,7 +53,7 @@ digraph statechart {
 	b -> c	 [label=" many_from [[:many_guard1, :many_guard2]]  ",
 		lp="236.14,122.17",
 		pos="e,151.5,98.852 115.51,145.79 122.79,140.34 129.85,134.14 135.59,127.35 140.3,121.76 144.23,115 147.4,108.35"];
-	"0024e893-f377-4461-9df6-65b95d76e7fa"	 [color=black,
+	"d0994e12-b365-4644-acaf-aa71a1247d69"	 [color=black,
 		fillcolor=black,
 		fixedsize=true,
 		height=0.30556,
@@ -62,5 +62,5 @@ digraph statechart {
 		shape=doublecircle,
 		style=filled,
 		width=0.30556];
-	c -> "0024e893-f377-4461-9df6-65b95d76e7fa"	 [pos="e,157.59,22.181 157.59,59.263 157.59,50.889 157.59,40.999 157.59,32.435"];
+	c -> "d0994e12-b365-4644-acaf-aa71a1247d69"	 [pos="e,157.59,22.181 157.59,59.263 157.59,50.889 157.59,40.999 157.59,32.435"];
 }
